move default tests to examples, fix missing tough-cookie dependency
