+++ performance/client/package.json
@@ -16,6 +16,7 @@
     "@types/js-yaml": "^3.12.5",
     "@types/node": "^13.13.15",
     "@types/node-cron": "^2.0.3",
+    "@types/tough-cookie": "^4.0.0",
     "@typescript-eslint/eslint-plugin": "^2.34.0",
     "@typescript-eslint/parser": "^2.34.0",
     "debug": "^3.1.0",
@@ -23,6 +24,7 @@
     "got": "^11.5.2",
     "jest": "^26.4.2",
     "js-yaml": "^3.14.0",
+    "tough-cookie": "^4.0.0",
     "typescript": "^3.9.7"
   },
   "scripts": {
