+++ performance/client/src/__tests__/examples/idle/index.ts
